Finish the "install" command.
