+++ README.md
@@ -32,7 +32,7 @@ design!
 2.  Provide **temporary root access via ADB**, can be revoked later.
         mia definition create my-phone --yes
         mia build my-phone
-        mia install my-phone --reboot
+        mia install my-phone
 
 3.  After the installation completed open F-Droid and update the applications
     list.
