+++ mia/helpers/android.py
@@ -2,6 +2,10 @@
 Provides helper function for interfacing with devices using ADB or getting
 information about the software or device.
 """
+
+import subprocess
+
+# Import custom helpers.
 from mia.helpers.utils import *
 
 
@@ -48,3 +52,58 @@ def get_cyanogenmod_release_version(recommended=True):
         return input_ask('Please provide a CM release version', None, True)
 
     return 'M12'
+
+
+def reboot_device(mode):
+    # Get the MIA handler singleton.
+    handler = MiaHandler()
+
+    if mode == 'bootloader' or mode == 'recovery':
+        if handler.args['--emulator']:
+            # Restart the emulator.
+            subprocess.call(['adb', '-e', 'reboot', mode])
+        else:
+            # Restart the device.
+            subprocess.call(['adb', 'reboot', mode])
+
+
+def set_open_recovery_script():
+    # Get the MIA handler singleton.
+    handler = MiaHandler()
+    # TODO: Check the md5sum of the files on the device, make sure they are OK.
+
+    # Push the open recovery script to the device.
+    script_path = os.path.join(handler.get_definition_path(),
+                               'other', 'openrecoveryscript')
+    push_file_to_device('file', script_path, '/sdcard/openrecoveryscript')
+
+    command = 'su root cp /sdcard/openrecoveryscript /cache/recovery/'
+    if handler.args['--emulator']:
+        # Run the command on the emulator.
+        subprocess.call(['adb', '-e', 'shell', command])
+    else:
+        # Run the command on the device.
+        subprocess.call(['adb', 'shell', command])
+
+
+def push_file_to_device(source_type, source, destination):
+    # Get the MIA handler singleton.
+    handler = MiaHandler()
+
+    # Get the file size.
+    file_size = os.path.getsize(source)
+
+    print('Pushing %s (%s) onto the device:\n - %s' %
+          (source_type, format_file_size(file_size), source))
+    print('Please wait...')
+
+    # TODO: Add progress bar.
+    # http://stackoverflow.com/questions/6595374/adb-push-pull-with-progress-bar
+    print('NOTE: Progress bar has not been implemented yet!')
+
+    if handler.args['--emulator']:
+        # Push file to the emulator.
+        subprocess.call(['adb', '-e', 'push', source, destination])
+    else:
+        # Push file to the device.
+        subprocess.call(['adb', 'push', source, destination])
